First part of indexing.
Given a file, we can compute its document vector.
Given multiple document vectors, we can compute their vocabularies.

A document vector is a map from term to term frequency (the number of
times that term occurs within that document).

A vocabulary is a map from term to document frequency (the number of
documents containing that term).