With external help: more idiomatic Clojure.
A friend told me that some things could have been done in a more
idiomatic way, and he pointed me at the `into' function.  So I tried out
what it does, and after having guessed that the hash-map/mapcat
combination was so-so.

I'm still not satisfied.  I'm thinking where `map' returns a seq of the
return values of the function, there must be a way to have something
that returns a (hash?) map of the return values of the function...