First try using `reduce', Practical Clojure.
Write a utility function that produces a map given a collection of keys
and a function.  (Each value in the map is given by applying the
function to the key.)

Use this utility function in the main code.

I'm not clear if this is more idiomatic Clojure than the previous code
using `into'.  But it was fun to write it :-)

Given COLL (collection of keys) and F (function), the first version was:

    (apply hash-map (mapcat (fn [x] [x (F x)]) COLL))

The second version was:

    (into {} (map (fn [x] [x (F x)]) COLL))

Now it's:

    (map-hash F COLL)

Well, of course with the provided implementation of `map-hash'.